requirements.txt
fastapi~=0.116.1
uvicorn~=0.35.0
pydantic~=2.11.7
llama-cpp-python
chromadb~=1.0.16

fastembed~=0.7.1
sse-starlette~=3.0.2
numpy~=2.3.2
h11~=0.16.0
pillow~=11.3.0
pip~=25.2
typing_extensions~=4.14.1
attrs~=25.3.0
distro~=1.9.0
protobuf~=6.31.1
filelock~=3.18.0
rsa~=4.9.1
pyasn1~=0.6.1
grpcio~=1.74.0
rich~=14.1.0
Pygments~=2.19.2
markdown-it-py~=3.0.0
requests~=2.32.4
colorama~=0.4.6
PyYAML~=6.0.2
zipp~=3.23.0
sniffio~=1.3.1
importlib_metadata~=8.7.0
pyproject_hooks~=1.2.0
build~=1.3.0
click~=8.2.1
httpcore~=1.0.9
idna~=3.10
httpx~=0.28.1
mdurl~=0.1.2
charset-normalizer~=3.4.2
setuptools~=80.3.1
overrides~=7.7.0
sympy~=1.14.0
mpmath~=1.3.0
typer~=0.16.0
shellingham~=1.5.4
fsspec~=2025.7.0
pyasn1_modules~=0.4.2
cachetools~=5.5.2
certifi~=2025.8.3
urllib3~=2.5.0
packaging~=25.0
MarkupSafe~=3.0.2
orjson~=3.11.1
PyPika~=0.48.9
backoff~=2.2.1
starlette~=0.47.2
pydantic_core~=2.33.2
websockets~=15.0.1
anyio~=4.10.0
posthog~=5.4.0
six~=1.17.0
python-dateutil~=2.9.0.post0
httptools~=0.6.4
watchfiles~=1.1.0
importlib_resources~=6.5.2
pybase64~=1.4.2
tenacity~=9.1.2
bcrypt~=4.3.0
onnxruntime~=1.22.1
jsonschema~=4.25.0
kubernetes~=33.1.0
mmh3~=5.2.0
tqdm~=4.67.1
oauthlib~=3.3.1
annotated-types~=0.7.0
tokenizers~=0.21.4
huggingface-hub~=0.34.4
loguru~=0.7.3
py_rust_stemmers~=0.1.5
diskcache~=5.6.3
Jinja2~=3.1.6
websocket-client~=1.8.0
pyreadline3~=3.5.4
referencing~=0.36.2
jsonschema-specifications~=2025.4.1
durationpy~=0.10
requests-oauthlib~=2.0.0
humanfriendly~=10.0
coloredlogs~=15.0.1
flatbuffers~=25.2.10
python-dotenv~=1.1.1